refactor: remove redundant game recovery state
