+++ src/main/java/com/codenames/codenames/backend/playingfield/GameManager.java
@@ -2,6 +2,8 @@ package com.codenames.codenames.backend.playingfield;
 
 import com.codenames.codenames.backend.clue.Clue;
 import com.codenames.codenames.backend.clue.ClueValidationService;
+import com.codenames.codenames.backend.serialization.CardDataTransferObject;
+import com.codenames.codenames.backend.serialization.GameStateDataTransferObject;
 import com.codenames.codenames.backend.utility.Color;
 import com.codenames.codenames.backend.utility.Role;
 import com.codenames.codenames.backend.utility.Team;
@@ -70,8 +72,9 @@ public class GameManager {
    * @param state bundled recovery state
    * @param clueValidationService clue validation service
    */
-  public GameManager(GameRecoveryState state, ClueValidationService clueValidationService) {
-    if (state.cards() == null || state.cards().isEmpty()) {
+  public GameManager(
+      GameStateDataTransferObject state, ClueValidationService clueValidationService) {
+    if (state.cardList() == null || state.cardList().isEmpty()) {
       throw new IllegalArgumentException("cards cannot be null or empty");
     }
     if (state.currentTurn() == null || state.currentPhase() == null) {
@@ -81,15 +84,34 @@ public class GameManager {
     this.currentTurn = state.currentTurn();
     this.currentPhase = state.currentPhase();
     this.winner = state.winner();
-    this.currentRedFound = state.currentRedFound();
-    this.currentBlueFound = state.currentBlueFound();
     this.remainingGuesses = state.remainingGuesses();
-    this.currentClue = state.currentClue();
+    this.currentClue =
+        state.currentClue() == null
+            ? null
+            : new Clue(state.currentClue().word(), state.currentClue().guessAmount());
     this.clueValidationService = clueValidationService;
 
-    this.redCards = countCardsByColor(state.cards(), Color.RED);
-    this.blueCards = countCardsByColor(state.cards(), Color.BLUE);
-    this.board = new Board(state.cards());
+    List<Card> cards = state.cardList().stream().map(this::toCard).toList();
+
+    this.currentRedFound = countGuessedByColor(cards, Color.RED);
+    this.currentBlueFound = countGuessedByColor(cards, Color.BLUE);
+
+    this.redCards = countCardsByColor(cards, Color.RED);
+    this.blueCards = countCardsByColor(cards, Color.BLUE);
+    this.board = new Board(cards);
+  }
+
+  private Card toCard(CardDataTransferObject cardDto) {
+    Card card = new Card(cardDto.word(), cardDto.color());
+    if (cardDto.isGuessed()) {
+      card.setIsGuessedTrue();
+    }
+    return card;
+  }
+
+  private int countGuessedByColor(List<Card> cards, Color color) {
+    return (int)
+        cards.stream().filter(card -> card.isGuessed() && card.getColor() == color).count();
   }
 
   /**
